+++ Classes/ViewHelpers/HtmlViewHelper.php
@@ -63,7 +63,7 @@ class HtmlViewHelper extends AbstractContentElementViewHelper
     {
         $html = $this->renderChildren();
         $htmlStyle = '';
-        $color = $this->convertHexToRGB($this->settings['generalText']['color']);
+        $color = $this->settingsConversionService->convertHexToRGB($this->settings['generalText']['color']);
         $padding = $this->settings['generalText']['padding'];
 
         if (!empty($this->arguments['styleSheet'])) {
@@ -86,7 +86,7 @@ class HtmlViewHelper extends AbstractContentElementViewHelper
         //reset settings to generalText
         $this->getPDF()->SetTextColor($color['R'], $color['G'], $color['B']);
         $this->getPDF()->SetFontSize($this->settings['generalText']['fontSize']);
-        $this->getPDF()->SetFont($this->settings['generalText']['fontFamily'], $this->convertToTcpdfFontStyle($this->settings['generalText']['fontStyle']));
+        $this->getPDF()->SetFont($this->settings['generalText']['fontFamily'], $this->settingsConversionService->convertFontStyle($this->settings['generalText']['fontStyle']));
         $this->getPDF()->setCellPaddings($padding['left'], $padding['top'], $padding['right'], $padding['bottom']);
 
         $this->getPDF()->writeHTML($htmlStyle . $html, true, false, true, false, '');